Warning fixes

+++ src/Color.h
@@ -22,7 +22,7 @@ public:
 		COLOR_LIMIT = 256,
 		COLOR_NONE = 0xFFFFFFFF,
 	};
-	cColor() { m_Color = COLOR_NONE;}
+	cColor() { m_Color = static_cast<unsigned int>(COLOR_NONE);}
 	cColor(unsigned char a_Red, unsigned char a_Green, unsigned char a_Blue) { SetColor(a_Red, a_Green, a_Blue); }
 
 	/// Returns whether the color is a valid color
@@ -50,7 +50,7 @@ public:
 	unsigned char GetBlue() const;
 
 	/// Resets the color
-	void Clear() { m_Color = COLOR_NONE; }
+	void Clear() { m_Color = static_cast<unsigned int>(COLOR_NONE); }
 	// tolua_end
 
 	unsigned int m_Color;
